% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_enhancers}
\alias{get_enhancers}
\title{Get a GRanges object with all the enhancers from Fantom.}
\usage{
get_enhancers()
}
\value{
A \code{GRanges} object representing every enhancers from Fantom in
  hg19.
}
\description{
Produce a GRanges object containing all the enhancers from Fantom but
without the expression values for every experiments.
}
\examples{
get_enhancers()
}
\seealso{
\code{\link{get_enhancers_tpm}}
}

