% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_fantom_enhancers_counts}
\alias{get_fantom_enhancers_counts}
\title{Get a GRanges object with the counts values for specific cell types}
\usage{
get_fantom_enhancers_counts(cell_lines = NULL, merge.FUN = NULL)
}
\arguments{
\item{cell_lines}{The cell line(s) to fetch. Must be a vector of character.
                  The function will look for the pattern (with \code{grepl})
                  to find case-insensitive match(es) with this param in the
                  Fantom's experiments sample names.

                  If \code{NULL}, this will return a \code{GRanges} object
                  with all the Fantom's experiments.

                  The metadata columns will be named using the
                  \code{cell_type} param. When \code{merge.FUN} is
                  \code{NULL} and there is multiple experiment for the same
                  cell line, a counter will be added after each name to
                  ensure all metadata colnames are unique (i.e.: A549_1,
                  A549_2, A549_3). If \code{NULL}, the experiment names will
                  be used for each metadata colnames.

                  Default: \code{NULL}.}

\item{merge.FUN}{A function to merge the TPM when there is more than one
                  column for the same cell type. Must take a vector of
                  numeric as input and returns a single numeric value.

                  Default: \code{NULL}}
}
\value{
A \code{GRanges} object with the metadata columns containing the
  expression value in counts for the requested cell line(s), all
  experiments expression values are returned if \code{cell_line} is
  \code{NULL}. If the cell line(s) is not found, an empty \code{GRanges} is
  returned.
}
\description{
Returns a GRanges with metadata columns corresponding to the requested
cell type enhancer expression (in counts).
}
\examples{
# To get the counts in lung
  get_fantom_enhancers_counts(cell_lines = "lung")

  # To get the counts in lung and heart
  get_fantom_enhancers_counts(cell_lines = c("lung", "heart"))

  # To get the counts lung and merge metadata columns by returning
  # their mean value
  get_fantom_enhancers_counts(cell_lines = "lung", merge.FUN = mean)
}
\seealso{
\code{\link{get_fantom_enhancers}}

\code{\link{get_fantom_enhancers_tpm}}

\code{\link{get_fantom_fantom_library_name}}
}

